move sample query tests
